HAL/vtable: introduce new HAL object - vtables
vtables are intended to aid implementation of runtime-loadable and
instantiable classses similar to C++.

At the core a vtable is named, typed, reference-counted pointer which can
be shared in a context (a context being RT, or within a userland HAL
component).

This commit adds HAL library support for HAL vtables.