tp/tp_shared.h: introduce the tp_shared_t comms structure
The idea is to enable shared access to state between using code and the tp
with a structure of pointers to HAL-typed values. The tp_shared_t structure
itself lives in HAL memory, so as to make it easy to switch any value
therein to a pin if so desired.

Using code in tp does not reference those items directly but rather through
wrapper functions in tp_shared.h .

While at it, the loose practices of typing motion and tp data items has
been firmed up - everything shared between tp and motion is now basic
HAL types; this requires some minor type changes in emcmot* structures.