hal/hal_lib: hal_lib becomes a component
So far hal_lib was very similar to a comp, but special-cased; for example,
ownership of the HAL shm segment (comps may own shm segments, and they are
automatically cleaned up on component exit). Also, there are HAL objects
ownership relations which suggest hooking them onto a 'root' HAL object,
for example userfuncts for newinst/delinst (functs are subordinate to a
comp, and cannot live without them).

Another example is HAL rings: as they persist their using component's lifecycle,
it makes sense to attach them to a root object so they are cleaned up
automatically on shutdown.

A small downside is that 'halcmd show comp' now shows new comps which were not
loaded by the user - a hal_lib instance for RT, and one for each userland
component. This could be mitigated in halcmd if needed.

The upside is hal_lib startup and shutdown becomes very regular, the rtapi_app_init
and rtapi_app_exit code is much simpler. To get around the startup 'chicken or egg'
problem, a new component type is used (TYPE_HALLIB) which signals to hal_xinit
how to special-case the bootstrap code.

userfuncts can be used for custom purposes, but are also used for newinst
and delinst, which is why they are attached to the RT hal_lib. So newinst/delinst
makes hal_lib eat its own dogfood ;)